Reorganized files into more appropriate places
