+++ public/index.html
@@ -18,7 +18,6 @@
       name="description"
       content="Create perfectly-formatted recital programs"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
