+++ src/App.js
@@ -1,7 +1,10 @@
+import Store from "./Store";
 import MainPage from "./pages/MainPage";
 
 export default function App() {
   return (
-    <MainPage />
+    <Store>
+      <MainPage />
+    </Store>
   );
 }
