+++ src/Reducer.js
@@ -1,0 +1,10 @@
+const Reducer = (state, action) => {
+    switch (action.type) {
+        case 'SET_ASLEEP':
+            return { ...state, asleep: action.payload };
+        default:
+            return state;
+    }
+};
+
+export default Reducer;
