+++ src/assets/icons/email-solid.png
