+++ src/assets/icons/trash-solid.png
