+++ src/assets/icons/upload-solid.png
